// src/app/models/AuctionList.ts
export class Car {
    id: number;
    registrationNumber: string;
    brand: string;
    model:  string;
    year: number;
    miles: number;
    price: number;
    color:  string;
    description:  string;
    sold: boolean;
    sellerId: number;
    soldDate: any;
    soldPrice: number;
    buyerId: number;
    createdDate: number;
    updatedDate: number;

    imagePath: string;
}
//
// export class Car{
//     id: number;
//     registrationNumber: string;
//     brand: string;
//     model:  string;
//     year: number;
//     miles: number;
//     price: number;
//     color:  string;
//     Description:  string;
//     sold: boolean;
//     sellerId: number;
//     soldDate: any;
//     soldPrice: number;
//     buyerId: number;
//     createdDate: number;
//     updatedDate: number;
// }


// id: number;
// title: string;
// description: string;
// imagePath: string;
// buyItNowPrice: number;
// bids: number;
// currentBidPrice: number;
// registrationDate: number;